workshops/workshop_filas/Kafka/pyproject.toml
[tool.poetry]
name = "kafka-workshop"
version = "0.1.0"
description = ""
authors = ["Lucas-Sobreira <lucas-andrade2011@live.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
confluent-kafka = "^2.6.0"
python-dotenv = "^1.0.1"
faker = "^30.8.0"
pandas = "^2.2.3"
sqlalchemy = "^2.0.36"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
